requirements.txt
crawlee==1.5.0
playwright==1.58.0
scrapling==0.4.2
pyyaml==6.0.3
markdownify==1.2.2
beautifulsoup4==4.14.3
lxml==6.0.2
tqdm==4.67.3
